Remove redundant test demos
