+++ demo/canvas-035.html
@@ -169,7 +169,11 @@
       <li>Create a Cell wrapper that hosts a Picture entity using a spritesheet as its asset, which can be animated and used as a pattern</li>
       <li>Create a Block entity that uses patterns for its fillStyle attribute</li>
       <li>Check that all patterns load and display as required</li>
-      <li>Check that all patterns are rendered appropriately when the display Block is moved and manipulated (note: patterns will not scale as the Block scales)</li>
+      <li>Check that all patterns are rendered appropriately when the display Block is moved and manipulated</li>
+      <ul>
+        <li>Cell-locked patterns will act like a static background covering the cell, appearing only where the entity fill or stroke appears; they do not move, scale or rotate with the entity</li>
+        <li>Entity-locked patterns will act as a fill or stroke on the entity which moves, scales and rotates with the entity</li> 
+      </ul>
       <li>Check that all patterns react appropriately as the 'repeat' attribute is updated (in real time)</li>
       <li>Check that all patterns react appropriately as each of the six pattern matrix values is manipulated (in real time)</li>
     </ul>
