+++ demo/canvas-209.js
@@ -1,5 +1,5 @@
 // # Demo Canvas 209
-// EnhancedLabel - clone entity, drag-drop; shadows, gradients and patterns
+// EnhancedLabel entity - keyboard navigation; hit tests
 
 // [Run code](../../demo/canvas-209.html)
 import * as scrawl from '../source/scrawl.js';
@@ -17,333 +17,222 @@ const namespace = canvas.name;
 const name = (n) => `${namespace}-${n}`;
 
 
-// #### Styling
-// Create gradient and pattern
-scrawl.makeGradient({
-    name: name('linear-gradient'),
-    endX: '100%',
-
-    colors: [
-        [0, 'blue'],
-        [495, 'red'],
-        [500, 'yellow'],
-        [505, 'red'],
-        [999, 'green'],
-    ],
-    colorSpace: 'OKLAB',
-    precision: 5,
+// #### User interaction: accessibility - keyboard
+canvas.set({
+    includeInTabNavigation: true,
 });
 
-scrawl.makePattern({
+let startIndex = -1,
+    endIndex = -1,
+    cursorIndex = 0,
+    selectionInProgress = false;
 
-    name: name('brick-pattern'),
-    imageSource: 'img/brick.png',
-});
+const clearSelection = () => {
 
-
-// #### User interaction: drag-drop
-// Make an object to hold functions we'll use for group-based UI
-const setCursorTo = {
-
-    auto: () => {
-        canvas.set({
-            css: {
-                cursor: 'auto',
-            },
-        });
-    },
-    pointer: () => {
-        canvas.set({
-            css: {
-                cursor: 'grab',
-            },
-        });
-    },
-    grabbing: () => {
-        canvas.set({
-            css: {
-                cursor: 'grabbing',
-            },
-        });
-    },
+    startIndex = -1;
+    endIndex = -1;
+    cursorIndex = 0;
+    selectionInProgress = false;
+    updateTextUnits(true);
 };
 
-// Create a drag group
-// + We define the EnhancedLabel entitys later - for now we just need their names
-const buttons = scrawl.makeGroup({ name: name('buttons-group') })
-    .addArtefacts(
-        name('Britain-label'),
-        name('Thailand-label'),
-        name('China-label'),
-        name('Egypt-label'),
-    );
+const updateSelection = () => {
 
-// Create two functions to do the `stampOrder` manipulation via the groups
-// + We create these button groups below
-const promoteStampOrders = (name) => {
-
-    // Promote the ones that match the name
-    if (name.includes('Britain')) britainGroup.set({ order: 9999 });
-    else if (name.includes('Thailand')) thailandGroup.set({ order: 9999 });
-    else if (name.includes('China')) chinaGroup.set({ order: 9999 });
-    else if (name.includes('Egypt')) egyptGroup.set({ order: 9999 });
+    startIndex = cursorIndex;
+    endIndex = cursorIndex;
+    selectionInProgress = !selectionInProgress;
+    updateTextUnits();
 };
 
-const demoteStampOrders = () => {
+const moveCursor = (direction) => {
 
-    // Demote them all!
-    britainGroup.set({ order: 1 });
-    thailandGroup.set({ order: 2 });
-    chinaGroup.set({ order: 3 });
-    egyptGroup.set({ order: 4 });
+    const maxIndex = mylabel.get('textUnits').length - 1;
+
+    switch (direction) {
+
+        case 'backwards' :
+
+            if (cursorIndex > 0) --cursorIndex;
+
+            break;
+
+        case 'forwards' :
+            if (cursorIndex < maxIndex) ++cursorIndex;
+            break;
+    }
+
+    if (selectionInProgress) endIndex = cursorIndex;
+
+    updateTextUnits();
 };
 
-// Use a state variable to keep track of which button is being dragged
-// + We also make use of this in the Display cycle report functionality
-let currentlyDragging = false;
+const copySelectionToConsole = () => {
 
-// With all the above in place, we can now create the drag-and-drop zone
-const userInteraction = scrawl.makeDragZone({
+    const units = mylabel.get('textUnits'),
+        maxIndex = units.length - 1;
+
+    if (startIndex >= 0 && startIndex <= maxIndex && endIndex >= 0 && endIndex <= maxIndex) {
+
+        const starts = (startIndex > endIndex) ? endIndex : startIndex,
+            ends = (startIndex > endIndex) ? startIndex : endIndex;
+
+        let text = '';
+
+        for (let i = starts; i <= ends; i++) {
+
+            text += units[i].chars;
+        }
+        console.log(`Text to copy: [${text}]`);
+    }
+    else console.log('Nothing to copy');
+}
+
+scrawl.makeKeyboardZone({
 
     zone: canvas,
-    collisionGroup: buttons,
-    endOn: ['up', 'leave'],
-    exposeCurrentArtefact: true,
-    preventTouchDefaultWhenDragging: true,
 
-    updateOnStart: () => {
-
-        // Invoking the function that `makeDragZone` returns gives us access to the entity being dragged (if any)
-/** @ts-expect-error */
-        currentlyDragging = userInteraction();
-
-        // Invoke the stampOrder promotion
-/** @ts-expect-error */
-        if (currentlyDragging) promoteStampOrders(currentlyDragging.artefact.name);
-
-        // Update the cursor to show we're dragging something
-        setCursorTo.grabbing();
+    ctrlOnly: {
+        c: () => copySelectionToConsole(),
     },
 
-    updateOnEnd: () => {
-
-        // Invoke the stampOrder demotion
-        demoteStampOrders();
-
-        // Set the state variable to false
-        currentlyDragging = false;
-
-        // Update the cursor - we're still hovering over the entity after dropping it
-        setCursorTo.pointer();
+    none: {
+        Backspace: () => clearSelection(),
+        Enter: () => updateSelection(),
+        ArrowLeft: () => moveCursor('backwards'),
+        ArrowUp: () => moveCursor('backwards'),
+        ArrowRight: () => moveCursor('forwards'),
+        ArrowDown: () => moveCursor('forwards'),
     },
 });
 
-
-// #### Groups
-// Each button gets its own group
-const britainGroup = scrawl.makeGroup({
-
-    name: name('Britain-group'),
-    order: 1,
-    host: canvas.base,
-
-    checkForEntityHover: true,
-    onEntityHover: setCursorTo.pointer,
-    onEntityNoHover: setCursorTo.auto,
-});
-
-const thailandGroup = britainGroup.clone({
-
-    name: name('Thailand-group'),
-    order: 2,
-});
-
-const chinaGroup = britainGroup.clone({
-
-    name: name('China-group'),
-    order: 3,
-});
-
-const egyptGroup = britainGroup.clone({
-
-    name: name('Egypt-group'),
-    order: 4,
-});
-
-
-// #### Entity generation
-// Helper functions
-const calculateInitialRoll = () => Math.random() * 360;
-
-const calculateInitialDeltaObject = () => {
-
-    let d = Math.random() - 0.5;
-    d += (d > 0) ? 0.2 : -0.2;
-
-    return { roll: d };
+const cursorHighlight = {
+    localStyle: {
+        includeHighlight: true,
+        highlightStyle: 'black',
+        fillStyle: 'yellow',
+    }
 };
 
-// Create the template entitys that will be used by our EnhancedLabel entitys
-scrawl.makeWheel({
+const selectionHighlight = {
+    localStyle: {
+        includeHighlight: true,
+        highlightStyle: 'orange',
+    }
+};
 
-    name: name('Britain-template'),
-    group: name('Britain-group'),
-    radius: 80,
-    start: ['30%', '25%'],
+const noHighlight = {
+    localStyle: null,
+};
+
+const updateTextUnits = (clear = false) => {
+
+    mylabel.setAllTextUnits(noHighlight);
+
+    if (!clear) {
+
+        const maxIndex = mylabel.get('textUnits').length - 1;
+
+        if (startIndex >= 0 && startIndex <= maxIndex && endIndex >= 0 && endIndex <= maxIndex) {
+
+            const starts = (startIndex > endIndex) ? endIndex : startIndex,
+                ends = (startIndex > endIndex) ? startIndex : endIndex;
+
+            for (let i = starts; i <= ends; i++) {
+
+                mylabel.setTextUnit(i, selectionHighlight);
+            }
+        }
+
+        if (cursorIndex >= 0 && cursorIndex <= maxIndex) mylabel.setTextUnit(cursorIndex, cursorHighlight);
+    }
+};
+
+
+// #### Create entitys
+const westernText = '<span class="underline">Lorem</span> ipsum <b>dolor sit</b> amet, con&shy;sectetur 😀 adi&shy;piscing &eacute;lit, sed <s>do eius-mod</s> <u>tempoj yn&shy;figizqunt</u> ut <span class="stamp-outlined">labore et dolore</span> <span class="green-highlight">magna aliqua.</span> Ut enim ad <span class="bold">minim veniam,</span> quis <span class="letter-spaced">nostrud</span> exercit-ation <span class="strike">ullamco laboris</span> nisi ut aliquip ex ea <span class="make-monospace">"commodo"</span> consequat. Duis <em>(aute irure d&ouml;lor)</em> in reprehenderit 🤖&icirc;n <i>voluptate</i> velit &copy;2024 <i>esse &lt;cillum&gt; <b>dolore</b> eu fug🎻iat nulla</i> pariatur. <span class="red">Excepteur sint</span> occaecat &iexcl;cupidatat! <strong>non proident,</strong> <span class="word-spaced">sunt in culpa qui</span> offici&thorn;a deserunt <span class="make-bigger"><span class="green-highlight">mollit</span> anim</span> id est laborum.';
+
+scrawl.makeSpiral({
+    name: name('spiral-track'),
+    strokeStyle: 'rgb(0 0 0 / 0.2)',
+    method: 'draw',
+    start: ['center', 'center'],
     handle: ['center', 'center'],
-    roll: calculateInitialRoll(),
-    method: 'none',
-
-}).clone({
-
-    name: name('Thailand-template'),
-    group: name('Thailand-group'),
-    roll: calculateInitialRoll(),
-    startX: '70%',
-
-}).clone({
-
-    name: name('China-template'),
-    group: name('China-group'),
-    roll: calculateInitialRoll(),
-    startY: '75%',
-
-}).clone({
-
-    name: name('Egypt-template'),
-    group: name('Egypt-group'),
-    roll: calculateInitialRoll(),
-    startX: '30%',
-
-// Create the background entitys
-}).clone({
-
-    name: name('Britain-background'),
-    group: name('Britain-group'),
-    pivot: name('Britain-template'),
-    lockTo: 'pivot',
-
-    radius: 90,
-
-    strokeStyle: 'rgb(0 0 150)',
-    fillStyle: 'rgb(200 200 255)',
-    lineWidth: 5,
-    method: 'fillThenDraw',
-
-}).clone({
-
-    name: name('Thailand-background'),
-    group: name('Thailand-group'),
-    pivot: name('Thailand-template'),
-    strokeStyle: 'rgb(150 150 0)',
-    fillStyle: 'rgb(255 255 200)',
-
-}).clone({
-
-    name: name('China-background'),
-    group: name('China-group'),
-    pivot: name('China-template'),
-    strokeStyle: 'rgb(150 0 0)',
-    fillStyle: 'rgb(255 200 200)',
-
-}).clone({
-
-    name: name('Egypt-background'),
-    group: name('Egypt-group'),
-    pivot: name('Egypt-template'),
-    strokeStyle: 'rgb(0 150 0)',
-    fillStyle: 'rgb(200 255 200)',
+    loops: 5,
+    loopIncrement: 50,
+    drawFromLoop: 2,
+    scaleOutline: false,
+    useAsPath: true,
 });
 
+const mylabel = scrawl.makeEnhancedLabel({
 
-// Create the EnhancedLabel entitys
-// + We finesse the word sizes and vertical positioning (including the underline) using CSS
-scrawl.makeEnhancedLabel({
+    name: name('my-label'),
+    fontString: '16px "Roboto Sans"',
+    text: westernText,
 
-    name: name('Britain-label'),
-    group: name('Britain-group'),
-    layoutTemplate: name('Britain-template'),
-
-    fontString: 'bold 24px "Roboto Serif',
-    text: '<p class="welcome-text">Welcome to <span class="country-name">Britain</span></p>',
-    startTextOnLine: 1,
-
-    delta: calculateInitialDeltaObject(),
-    noDeltaUpdates: true,
+    layoutTemplate: name('spiral-track'),
+    useLayoutTemplateAsPath: true,
 
     textHandle: ['center', 'alphabetic'],
-    lineSpacing: 2,
-    lineAdjustment: 28,
-    letterSpacing: '2px',
-    wordSpacing: '2px',
 
-    fillStyle: `rgb(220 170 140)`,
+    breakTextOnSpaces: false,
 
-    shadowOffsetX: 2,
-    shadowOffsetY: 2,
-    shadowBlur: 1,
-    shadowColor: 'black',
+    checkHitUseTemplate: false,
 
-    globalAlpha: 1,
-    globalCompositeOperation: 'source-over',
-
-}).clone({
-
-    name: name('Thailand-label'),
-    group: name('Thailand-group'),
-    layoutTemplate: name('Thailand-template'),
-
-    fontString: 'bold 24px "Noto Thai Sans',
-    text: '<p class="welcome-text">ยินดีต้อนรับสู่ <span class="country-name">ประเทศไทย</span></p>',
-
-    delta: calculateInitialDeltaObject(),
-
-}).clone({
-
-    name: name('China-label'),
-    group: name('China-group'),
-    layoutTemplate: name('China-template'),
-
-    fontString: 'bold 24px "Noto Chinese Simple Sans',
-    text: '<p class="welcome-text">欢迎来到 <span class="country-name">中国</span></p>',
-
-    delta: calculateInitialDeltaObject(),
-
-}).clone({
-
-    name: name('Egypt-label'),
-    group: name('Egypt-group'),
-    layoutTemplate: name('Egypt-template'),
-
-    fontString: 'bold 24px "Noto Arabic Sans',
-    text: '<p class="welcome-text">مرحبا بكم فى <span class="country-name">مصر</span></p>',
-
-    delta: calculateInitialDeltaObject(),
-
-    direction: 'rtl',
-    letterSpacing: '0px',
-    wordSpacing: '4px',
+    delta: {
+        pathPosition: 0.0005,
+    },
+    noDeltaUpdates: true,
 });
+
+
+// #### User interaction: mouse/touch
+const checkMouseHover = () => {
+
+    const hit = mylabel.checkHit(canvas.here);
+
+    if (hit && typeof hit !== 'boolean' && hit.index != null) {
+
+        if (selectionInProgress) endIndex = hit.index;
+        else cursorIndex = hit.index;
+
+        updateTextUnits();
+    }
+};
+
+scrawl.addListener('down', () => {
+
+    updateTextUnits(true);
+
+    const hit = mylabel.checkHit(canvas.here);
+
+    if (hit && typeof hit !== 'boolean' && hit.index != null) {
+
+        cursorIndex = hit.index;
+        startIndex = cursorIndex;
+        endIndex = cursorIndex;
+        selectionInProgress = true;
+    }
+    else {
+
+        cursorIndex = -1;
+        startIndex = -1;
+        endIndex = -1;
+        selectionInProgress = false;
+    }
+
+}, canvas.domElement);
+
+scrawl.addListener(['up', 'leave'], () => {
+
+    selectionInProgress = false;
+
+}, canvas.domElement);
 
 
 // #### Scene animation
 // Function to display frames-per-second data, and other information relevant to the demo
-const report = reportSpeed('#reportmessage', function () {
-
-/** @ts-expect-error */
-    const dragReport = `Currently dragging: ${(typeof currentlyDragging !== 'boolean' && currentlyDragging) ? currentlyDragging.artefact.name : 'nothing'}`;
-
-    let fontReadout = `
-`;
-    document.fonts.forEach(k => {
-        if (k.status === 'loaded') fontReadout +=(`    ${k.family} ${k.weight} ${k.style}\n`)
-    });
-
-    return `
-${dragReport}
-
-Loaded fonts:${fontReadout}`;
-});
+const report = reportSpeed('#reportmessage');
 
 
 // Create the Display cycle animation
@@ -351,53 +240,107 @@ scrawl.makeRender({
 
     name: name('animation'),
     target: canvas,
-    commence: () => canvas.checkHover(),
+    commence: checkMouseHover,
     afterShow: report,
 });
 
 
-// #### User controls
+// #### User form interaction
+// Setup form
+const dom = scrawl.initializeDomInputs([
+    ['input', 'alignment', '0'],
+    ['input', 'letterSpacing', '0'],
+    ['input', 'roll', '0'],
+    ['input', 'scale', '1'],
+    ['input', 'wordSpacing', '0'],
+    ['select', 'animation', 0],
+    ['select', 'breakTextOnSpaces', 1],
+    ['select', 'font', 0],
+    ['select', 'textUnitFlow', 0],
+    ['select', 'flipReverse', 0],
+    ['select', 'flipUpend', 0],
+    ['select', 'alignTextUnitsToPath', 1],
+]);
+
+
+const updateAnimation = (event) => {
+
+    const val = event.target.value;
+
+    if (val) mylabel.set({ noDeltaUpdates: false });
+    else mylabel.set({ noDeltaUpdates: true });
+};
+scrawl.addNativeListener('change', (e) => updateAnimation(e), dom.animation);
+
+
+const updateFont = (event) => {
+
+    const font = event.target.value;
+
+    if (font) {
+
+        switch (font) {
+
+            case 'roboto-serif' :
+                mylabel.set({
+                    fontString: '16px "Roboto Serif"',
+                    text: westernText,
+                    direction: 'ltr',
+                });
+                break;
+
+            case 'noto-hebrew-serif' :
+                mylabel.set({
+                    fontString: '16px "Noto Hebrew Serif"',
+                    text: 'כל אדם זכאי לחירות הדעה והבטוי, לרבות החירות להחיק בדעות ללא כל הפרעה, ולבקש ידיעות ודעות, ולקבלן ולמסרן בכל הדרכים וללא סייגי גבולות כל אדם, כחבר החברה, זכאי לבטחון סוציאלי וזכאי לתבוע שהזכויות הכלכליות הסוציאליות והתרבותיות, שהן חיוניות לכבודו כאדם ולהתפתחות החופשית של אישיותו, יובטחו במשמץ לאומי ובשיתוף פעולה בינלאומי בהתאם לארגונה ולאוצרותיה של המדינה כל אדם זכאי למנוחה ולפנאי',
+                    direction: 'rtl',
+                });
+                break;
+
+            case 'noto-japanese-sans' :
+                mylabel.set({
+                    fontString: '16px "Noto Japanese Sans"',
+                    text: '人類社会のすべての構成員の固有の尊厳と平等で譲ることのできない権利とを承認することは&#x2060;、世界における自由&#x2060;、正義及び平和の基礎であるので&#x2060;、 人権の無視及び軽侮が&#x2060;、人類の良心を踏みにじった野蛮行為をもたらし&#x2060;、言論及び信仰の自由が受けられ&#x2060;、恐怖及び欠乏のない世界の到来が&#x2060;、一般の人々の最高の願望として宣言されたので&#x2060;、 人間が専制と圧迫とに対する最後の手段として反逆に訴えることがないようにするためには&#x2060;、法の支配によって人権を保護することが肝要であるので&#x2060;、 諸国間の友好関係の発展を促進することが肝要であるので&#x2060;、',
+                    direction: 'ltr',
+                });
+                break;
+        }
+    }
+};
+scrawl.addNativeListener('change', (e) => updateFont(e), dom.font);
+
+
 scrawl.makeUpdater({
 
     event: ['input', 'change'],
     origin: '.controlItem',
 
-    target: buttons,
+    target: mylabel,
 
     useNativeListener: true,
     preventDefault: true,
 
     updates: {
-        lockFillStyleToEntity: ['lockFillStyleToEntity', 'boolean'],
-        noDeltaUpdates: ['noDeltaUpdates', 'boolean'],
-        ['text-fillStyle']: ['fillStyle', 'raw'],
-        ['underline-fillStyle']: ['underlineStyle', 'raw'],
-        globalCompositeOperation: ['globalCompositeOperation', 'raw'],
-        globalAlpha: ['globalAlpha', 'float'],
+
+        letterSpacing: ['letterSpacing', 'px'],
+        wordSpacing: ['wordSpacing', 'px'],
+
+        roll: ['roll', 'float'],
+        scale: ['scale', 'float'],
+
+        textUnitFlow: ['textUnitFlow', 'raw'],
+
+        alignment: ['alignment', 'float'],
+
+        alignTextUnitsToPath: ['alignTextUnitsToPath', 'boolean'],
+        breakTextOnSpaces: ['breakTextOnSpaces', 'boolean'],
+
+        flipReverse: ['flipReverse', 'boolean'],
+        flipUpend: ['flipUpend', 'boolean'],
     },
+
+    callback: clearSelection,
 });
-
-
-// Setup form
-scrawl.initializeDomInputs([
-    ['input', 'globalAlpha', '1'],
-    ['select', 'text-fillStyle', 0],
-    ['select', 'underline-fillStyle', 0],
-    ['select', 'lockFillStyleToEntity', 0],
-    ['select', 'globalCompositeOperation', 0],
-]);
-// /** @ts-expect-error */
-// document.querySelector('#text-fillStyle').options.selectedIndex = 0;
-// /** @ts-expect-error */
-// document.querySelector('#underline-fillStyle').options.selectedIndex = 0;
-// /** @ts-expect-error */
-// document.querySelector('#lockFillStyleToEntity').options.selectedIndex = 0;
-// /** @ts-expect-error */
-// document.querySelector('#noDeltaUpdates').options.selectedIndex = 0;
-// /** @ts-expect-error */
-// document.querySelector('#globalCompositeOperation').options.selectedIndex = 0;
-// /** @ts-expect-error */
-// document.querySelector('#globalAlpha').value = 1;
 
 
 // #### Development and testing
