+++ demo/canvas-210.html
@@ -30,7 +30,7 @@
 <body>
   <div id="test-navigation">
     <a href="canvas-209.html">&larr; Previous</a>
-    <a href="canvas-211.html">Next &rarr;</a>
+    <a href="filters-001.html">Next &rarr;</a>
   </div>
   <h1><a href="index.html">Scrawl-canvas v8</a> - Canvas test 210</h1>
   <h2>EnhancedLabel - TextUnit dynamic manipulation</h2>
