+++ demo/filters-001.html
@@ -54,7 +54,7 @@
 
 <body>
   <div id="test-navigation">
-    <a href="canvas-212.html">&larr; Previous</a>
+    <a href="canvas-210.html">&larr; Previous</a>
     <a href="filters-002.html">Next &rarr;</a>
   </div>
   <h1><a href="index.html">Scrawl-canvas v8</a> - Filters test 001</h1>
