+++ demo/index.html
@@ -756,27 +756,14 @@
       <div class="no-padding"><img src="thumbs/canvas-209.webp" /></div>
       <div>
         <a href="canvas-209.html">Canvas 209</a>
-        EnhancedLabel - clone entity, drag-drop; shadows, gradients and patterns
-        <br /><span class="responsive">Drag zones</span>
+        EnhancedLabel entity - keyboard navigation; hit tests
+        <br /><span class="responsive">Keyboard zones</span>
       </div>
       
       <div class="no-padding"><img src="thumbs/canvas-210.webp" /></div>
       <div>
         <a href="canvas-210.html">Canvas 210</a>
         EnhancedLabel - TextUnit dynamic manipulation
-      </div>
-      
-      <div class="no-padding"><img src="thumbs/canvas-211.webp" /></div>
-      <div>
-        <a href="canvas-211.html">Canvas 211</a>
-        EnhancedLabel entity - keyboard navigation; hit tests
-        <br /><span class="responsive">Keyboard zones</span>
-      </div>
-      
-      <div class="no-padding"><img src="thumbs/canvas-212.webp" /></div>
-      <div>
-        <a href="canvas-212.html">Canvas 212</a>
-        EnhancedLabel entity - updating font parts; experiments with variable font
       </div>
       
     </div>
