+++ docs/index.html
@@ -568,10 +568,8 @@
     <li><a href="demo/canvas-206.html">Canvas-206</a> EnhancedLabel entity - basic multiline text</li>
     <li><a href="demo/canvas-207.html">Canvas-207</a> EnhancedLabel entity - styling multiline text</li>
     <li><a href="demo/canvas-208.html">Canvas-208</a> EnhancedLabel entity - text along a path</li>
-    <li><a href="demo/canvas-209.html">Canvas-209</a> EnhancedLabel - clone entity, drag-drop; shadows, gradients and patterns</li>
+    <li><a href="demo/canvas-209.html">Canvas-209</a> EnhancedLabel entity - keyboard navigation; hit tests</li>
     <li><a href="demo/canvas-210.html">Canvas-210</a> EnhancedLabel - TextUnit dynamic manipulation</li>
-    <li><a href="demo/canvas-211.html">Canvas-211</a> EnhancedLabel entity - keyboard navigation; hit tests</li>
-    <li><a href="demo/canvas-212.html">Canvas-212</a> EnhancedLabel entity - updating font parts; experiments with variable font</li>
   </ul>
 
   <h4>Filters demos</h4>
